UI: Allow theme to be toggled between light and dark
Currently when the app is launched the theme of the app is set to
light and it cannot be changed. This commit adds support to toggle the
theme between light and dark in the settings tab. However, currently
the chosen theme will not be saved and the user will need to change the
theme when the app is relaunched.